+++ CHANGELOG.md
@@ -1,15 +1,25 @@
 # Changelog
 
+## [2.7.0] - 20 May 2022
+
+### Added
+- Add possibility to set the app's notification channel's ID, name, and description on `Android` via `notificationChannelId`, `notificationChannelName`, and `notificationChannelDescription` settings keys in `selligent.json`
+
+### Changed
+- Add version 3.9.0 of the Android SDK to `src/android/frameworks`
+- Add version 2.7.4 of the iOS SDK static library to `src/ios/frameworks`
+- Change used native SDK version numbers to `3.9.0` and `2.7.4` for Android and iOS respectively in README
+
 ## [2.6.0] - 24 February 2022
 
 ### Added
-- Re-added the geolocation functionalities for Android and iOS
-- Re-added geolocation documentation to README
+- Re-add the geolocation functionalities for Android and iOS
+- Re-add geolocation documentation to README
 
 ### Changed
-- Added version 3.8.1 of the Android SDK to `src/android/frameworks`
-- Added version 2.7.2 of the iOS SDK static library to `src/ios/frameworks`
-- Updated the SDK versions to `3.8.1` and `2.7.2` for Android and iOS respectively in README
+- Add version 3.8.1 of the Android SDK to `src/android/frameworks`
+- Add version 2.7.2 of the iOS SDK static library to `src/ios/frameworks`
+- Update the SDK versions to `3.8.1` and `2.7.2` for Android and iOS respectively in README
 - Change this project's version number to `2.6.0` in `package.json`, `package-lock.json`, and `plugin.xml`
 
 ### Fixed